comparisons in dedicated module
